requirements.txt
markdown
pdfminer.six

langchain==0.1.2
langchain-core==0.1.19
langchain-community==0.0.18

InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
huggingface-hub==0.8.1
faiss-cpu==1.9.0

llama-cpp-python==0.3.1

# It is need rust compiler version less than 1.73 and export RUSTUP_TOOLCHAIN=_RUST_VERSION_
